feat(postgresql): Add support for overwrite-deletestale
