+++ plugins/source/gcp/go.mod
@@ -31,6 +31,8 @@ require (
 	google.golang.org/protobuf v1.28.1
 )
 
+replace github.com/cloudquery/plugin-sdk => ../../../../plugin-sdk
+
 require (
 	cloud.google.com/go v0.104.0 // indirect
 	cloud.google.com/go/iam v0.4.0 // indirect
